Split the classes into their own files. Add folders where needed.
