+++ package.json
@@ -21,14 +21,14 @@
   "activationEvents": [
     "onLanguage:markdown"
   ],
-  "main": "./out/src/extension",
+  "main": "./out/src/extension/extension",
   "contributes": { },
   "capabilities": {
     "documentFormattingProvider" : "true"
   },
   "scripts": {
     "vscode:prepublish": "tsc -p ./",
-    "compile": "tsc -watch -p ./",
+    "compile": "tsc -p ./",
     "postinstall": "node ./node_modules/vscode/bin/install",
     "test": "node ./node_modules/vscode/bin/test"
   },
