+++ src/column/columnPositioning.ts
@@ -1,0 +1,6 @@
+export enum ColumnPositioning {
+    Unkown,
+    First,
+    Middle,
+    Last
+}
