+++ test/extension/extension.test.ts
@@ -6,7 +6,6 @@
 // The module 'assert' provides assertion methods from node
 import * as assert from 'assert';
 import * as vscode from 'vscode';
-import * as myExtension from '../src/extension';
 
 // Defines a Mocha test suite to group tests of similar kind together
 suite("Extension Tests", () => {
